test(agents): split session loop lifecycle coverage
